Add resource to spec
