+++ src/controller_examples/rabbitmq_controller/exec/resource/role.rs
@@ -12,7 +12,7 @@ use crate::pervasive_ext::string_map::StringMap;
 use crate::pervasive_ext::string_view::*;
 use crate::rabbitmq_controller::common::*;
 use crate::rabbitmq_controller::exec::types::*;
-use crate::rabbitmq_controller::spec::reconciler as rabbitmq_spec;
+use crate::rabbitmq_controller::spec::resource as spec_resource;
 use crate::reconciler::exec::{io::*, reconciler::*};
 use vstd::prelude::*;
 use vstd::seq_lib::*;
@@ -25,7 +25,7 @@ pub fn update_role(rabbitmq: &RabbitmqCluster, found_role: Role) -> (role: Role)
         rabbitmq@.metadata.name.is_Some(),
         rabbitmq@.metadata.namespace.is_Some(),
     ensures
-        role@ == rabbitmq_spec::update_role(rabbitmq@, found_role@),
+        role@ == spec_resource::update_role(rabbitmq@, found_role@),
 {
     let mut role = found_role.clone();
     let made_role = make_role(rabbitmq);
@@ -46,7 +46,7 @@ pub fn make_policy_rules(rabbitmq: &RabbitmqCluster) -> (rules: Vec<PolicyRule>)
         rabbitmq@.metadata.name.is_Some(),
         rabbitmq@.metadata.namespace.is_Some(),
     ensures
-        rules@.map_values(|r: PolicyRule| r@) == rabbitmq_spec::make_role(rabbitmq@).policy_rules.get_Some_0(),
+        rules@.map_values(|r: PolicyRule| r@) == spec_resource::make_role(rabbitmq@).policy_rules.get_Some_0(),
 {
     let mut rules = Vec::new();
     rules.push({
@@ -57,7 +57,7 @@ pub fn make_policy_rules(rabbitmq: &RabbitmqCluster) -> (rules: Vec<PolicyRule>)
             proof{
                 assert_seqs_equal!(
                     api_groups@.map_values(|p: String| p@),
-                    rabbitmq_spec::make_role(rabbitmq@).policy_rules.get_Some_0()[0].api_groups.get_Some_0()
+                    spec_resource::make_role(rabbitmq@).policy_rules.get_Some_0()[0].api_groups.get_Some_0()
                 );
             }
             api_groups
@@ -68,7 +68,7 @@ pub fn make_policy_rules(rabbitmq: &RabbitmqCluster) -> (rules: Vec<PolicyRule>)
             proof{
                 assert_seqs_equal!(
                     resources@.map_values(|p: String| p@),
-                    rabbitmq_spec::make_role(rabbitmq@).policy_rules.get_Some_0()[0].resources.get_Some_0()
+                    spec_resource::make_role(rabbitmq@).policy_rules.get_Some_0()[0].resources.get_Some_0()
                 );
             }
             resources
@@ -79,7 +79,7 @@ pub fn make_policy_rules(rabbitmq: &RabbitmqCluster) -> (rules: Vec<PolicyRule>)
             proof{
                 assert_seqs_equal!(
                     verbs@.map_values(|p: String| p@),
-                    rabbitmq_spec::make_role(rabbitmq@).policy_rules.get_Some_0()[0].verbs
+                    spec_resource::make_role(rabbitmq@).policy_rules.get_Some_0()[0].verbs
                 );
             }
             verbs
@@ -94,7 +94,7 @@ pub fn make_policy_rules(rabbitmq: &RabbitmqCluster) -> (rules: Vec<PolicyRule>)
             proof{
                 assert_seqs_equal!(
                     api_groups@.map_values(|p: String| p@),
-                    rabbitmq_spec::make_role(rabbitmq@).policy_rules.get_Some_0()[1].api_groups.get_Some_0()
+                    spec_resource::make_role(rabbitmq@).policy_rules.get_Some_0()[1].api_groups.get_Some_0()
                 );
             }
             api_groups
@@ -105,7 +105,7 @@ pub fn make_policy_rules(rabbitmq: &RabbitmqCluster) -> (rules: Vec<PolicyRule>)
             proof{
                 assert_seqs_equal!(
                     resources@.map_values(|p: String| p@),
-                    rabbitmq_spec::make_role(rabbitmq@).policy_rules.get_Some_0()[1].resources.get_Some_0()
+                    spec_resource::make_role(rabbitmq@).policy_rules.get_Some_0()[1].resources.get_Some_0()
                 );
             }
             resources
@@ -116,7 +116,7 @@ pub fn make_policy_rules(rabbitmq: &RabbitmqCluster) -> (rules: Vec<PolicyRule>)
             proof{
                 assert_seqs_equal!(
                     verbs@.map_values(|p: String| p@),
-                    rabbitmq_spec::make_role(rabbitmq@).policy_rules.get_Some_0()[1].verbs
+                    spec_resource::make_role(rabbitmq@).policy_rules.get_Some_0()[1].verbs
                 );
             }
             verbs
@@ -126,7 +126,7 @@ pub fn make_policy_rules(rabbitmq: &RabbitmqCluster) -> (rules: Vec<PolicyRule>)
     proof{
         assert_seqs_equal!(
             rules@.map_values(|p: PolicyRule| p@),
-            rabbitmq_spec::make_role(rabbitmq@).policy_rules.get_Some_0()
+            spec_resource::make_role(rabbitmq@).policy_rules.get_Some_0()
         );
     }
     rules
@@ -137,7 +137,7 @@ pub fn make_role(rabbitmq: &RabbitmqCluster) -> (role: Role)
         rabbitmq@.metadata.name.is_Some(),
         rabbitmq@.metadata.namespace.is_Some(),
     ensures
-        role@ == rabbitmq_spec::make_role(rabbitmq@),
+        role@ == spec_resource::make_role(rabbitmq@),
 {
     let mut role = Role::default();
     role.set_metadata({
