+++ src/controller_examples/rabbitmq_controller/exec/resource/stateful_set.rs
@@ -12,7 +12,7 @@ use crate::pervasive_ext::string_map::StringMap;
 use crate::pervasive_ext::string_view::*;
 use crate::rabbitmq_controller::common::*;
 use crate::rabbitmq_controller::exec::types::*;
-use crate::rabbitmq_controller::spec::reconciler as rabbitmq_spec;
+use crate::rabbitmq_controller::spec::resource as spec_resource;
 use crate::reconciler::exec::{io::*, reconciler::*};
 use vstd::prelude::*;
 use vstd::seq_lib::*;
@@ -26,7 +26,7 @@ pub fn update_stateful_set(rabbitmq: &RabbitmqCluster, found_stateful_set: State
         rabbitmq@.metadata.namespace.is_Some(),
         found_stateful_set@.spec.is_Some(),
     ensures
-        stateful_set@ == rabbitmq_spec::update_stateful_set(rabbitmq@, found_stateful_set@, config_map_rv@),
+        stateful_set@ == spec_resource::update_stateful_set(rabbitmq@, found_stateful_set@, config_map_rv@),
 {
     let made_sts = make_stateful_set(rabbitmq, config_map_rv);
 
@@ -57,7 +57,7 @@ pub fn update_stateful_set(rabbitmq: &RabbitmqCluster, found_stateful_set: State
 
 pub fn sts_restart_annotation() -> (anno: String)
     ensures
-        anno@ == rabbitmq_spec::sts_restart_annotation(),
+        anno@ == spec_resource::sts_restart_annotation(),
 {
     new_strlit("anvil.dev/lastRestartAt").to_string()
 }
@@ -67,7 +67,7 @@ pub fn make_stateful_set_name(rabbitmq: &RabbitmqCluster) -> (name: String)
         rabbitmq@.metadata.name.is_Some(),
         rabbitmq@.metadata.namespace.is_Some(),
     ensures
-        name@ == rabbitmq_spec::make_stateful_set_name(rabbitmq@.metadata.name.get_Some_0()),
+        name@ == spec_resource::make_stateful_set_name(rabbitmq@.metadata.name.get_Some_0()),
 {
     rabbitmq.name().unwrap().concat(new_strlit("-server"))
 }
@@ -77,7 +77,7 @@ pub fn make_stateful_set(rabbitmq: &RabbitmqCluster, config_map_rv: &String) -> 
         rabbitmq@.metadata.name.is_Some(),
         rabbitmq@.metadata.namespace.is_Some(),
     ensures
-        stateful_set@ == rabbitmq_spec::make_stateful_set(rabbitmq@, config_map_rv@),
+        stateful_set@ == spec_resource::make_stateful_set(rabbitmq@, config_map_rv@),
 {
     let mut stateful_set = StatefulSet::default();
     stateful_set.set_metadata({
@@ -112,7 +112,7 @@ pub fn make_stateful_set(rabbitmq: &RabbitmqCluster, config_map_rv: &String) -> 
                 proof {
                     assert_seqs_equal!(
                         empty_pvc@.map_values(|pvc: PersistentVolumeClaim| pvc@),
-                        rabbitmq_spec::make_stateful_set(rabbitmq@, config_map_rv@).spec.get_Some_0().volume_claim_templates.get_Some_0()
+                        spec_resource::make_stateful_set(rabbitmq@, config_map_rv@).spec.get_Some_0().volume_claim_templates.get_Some_0()
                     );
                 }
                 empty_pvc
@@ -139,7 +139,7 @@ pub fn make_stateful_set(rabbitmq: &RabbitmqCluster, config_map_rv: &String) -> 
                             proof {
                                 assert_seqs_equal!(
                                     access_modes@.map_values(|mode: String| mode@),
-                                    rabbitmq_spec::make_stateful_set(rabbitmq@, config_map_rv@)
+                                    spec_resource::make_stateful_set(rabbitmq@, config_map_rv@)
                                         .spec.get_Some_0().volume_claim_templates.get_Some_0()[0]
                                         .spec.get_Some_0().access_modes.get_Some_0()
                                 );
@@ -164,7 +164,7 @@ pub fn make_stateful_set(rabbitmq: &RabbitmqCluster, config_map_rv: &String) -> 
                 proof {
                     assert_seqs_equal!(
                         volume_claim_templates@.map_values(|pvc: PersistentVolumeClaim| pvc@),
-                        rabbitmq_spec::make_stateful_set(rabbitmq@, config_map_rv@).spec.get_Some_0().volume_claim_templates.get_Some_0()
+                        spec_resource::make_stateful_set(rabbitmq@, config_map_rv@).spec.get_Some_0().volume_claim_templates.get_Some_0()
                     );
                 }
                 volume_claim_templates
@@ -206,7 +206,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
         rabbitmq@.metadata.name.is_Some(),
         rabbitmq@.metadata.namespace.is_Some(),
     ensures
-        pod_spec@ == rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@),
+        pod_spec@ == spec_resource::make_rabbitmq_pod_spec(rabbitmq@),
 {
     let mut volumes = Vec::new();
     volumes.push({
@@ -248,7 +248,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
                             proof {
                                 assert_seqs_equal!(
                                     items@.map_values(|item: KeyToPath| item@),
-                                    rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@).volumes.get_Some_0()[1].projected.get_Some_0()
+                                    spec_resource::make_rabbitmq_pod_spec(rabbitmq@).volumes.get_Some_0()[1].projected.get_Some_0()
                                     .sources.get_Some_0()[0].config_map.get_Some_0().items.get_Some_0()
                                 );
                             }
@@ -274,7 +274,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
                             proof {
                                 assert_seqs_equal!(
                                     items@.map_values(|item: KeyToPath| item@),
-                                    rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@).volumes.get_Some_0()[1].projected.get_Some_0()
+                                    spec_resource::make_rabbitmq_pod_spec(rabbitmq@).volumes.get_Some_0()[1].projected.get_Some_0()
                                     .sources.get_Some_0()[1].secret.get_Some_0().items.get_Some_0()
                                 );
                             }
@@ -287,7 +287,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
                 proof {
                     assert_seqs_equal!(
                         sources@.map_values(|source: VolumeProjection| source@),
-                        rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@).volumes.get_Some_0()[1].projected.get_Some_0()
+                        spec_resource::make_rabbitmq_pod_spec(rabbitmq@).volumes.get_Some_0()[1].projected.get_Some_0()
                         .sources.get_Some_0()
                     );
                 }
@@ -339,7 +339,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
                 proof {
                     assert_seqs_equal!(
                         items@.map_values(|item: DownwardAPIVolumeFile| item@),
-                        rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@).volumes.get_Some_0()[5].downward_api.get_Some_0().items.get_Some_0()
+                        spec_resource::make_rabbitmq_pod_spec(rabbitmq@).volumes.get_Some_0()[5].downward_api.get_Some_0().items.get_Some_0()
                     );
                 }
                 items
@@ -359,7 +359,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
     proof {
         assert_seqs_equal!(
             volumes@.map_values(|vol: Volume| vol@),
-            rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@).volumes.get_Some_0()
+            spec_resource::make_rabbitmq_pod_spec(rabbitmq@).volumes.get_Some_0()
         );
     }
     let mut pod_spec = PodSpec::default();
@@ -442,7 +442,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
                 proof {
                     assert_seqs_equal!(
                         volume_mounts@.map_values(|volume_mount: VolumeMount| volume_mount@),
-                        rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@).init_containers.unwrap()[0].volume_mounts.get_Some_0()
+                        spec_resource::make_rabbitmq_pod_spec(rabbitmq@).init_containers.unwrap()[0].volume_mounts.get_Some_0()
                     );
                 }
                 volume_mounts
@@ -452,7 +452,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
         proof {
             assert_seqs_equal!(
                 containers@.map_values(|container: Container| container@),
-                rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@).init_containers.unwrap()
+                spec_resource::make_rabbitmq_pod_spec(rabbitmq@).init_containers.unwrap()
             );
         }
         containers
@@ -515,7 +515,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
                 proof {
                     assert_seqs_equal!(
                         volume_mounts@.map_values(|volume_mount: VolumeMount| volume_mount@),
-                        rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@).containers[0].volume_mounts.get_Some_0()
+                        spec_resource::make_rabbitmq_pod_spec(rabbitmq@).containers[0].volume_mounts.get_Some_0()
                     );
                 }
                 volume_mounts
@@ -529,7 +529,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
                 proof {
                     assert_seqs_equal!(
                         ports@.map_values(|port: ContainerPort| port@),
-                        rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@).containers[0].ports.get_Some_0()
+                        spec_resource::make_rabbitmq_pod_spec(rabbitmq@).containers[0].ports.get_Some_0()
                     );
                 }
 
@@ -554,7 +554,7 @@ pub fn make_rabbitmq_pod_spec(rabbitmq: &RabbitmqCluster) -> (pod_spec: PodSpec)
         proof {
             assert_seqs_equal!(
                 containers@.map_values(|container: Container| container@),
-                rabbitmq_spec::make_rabbitmq_pod_spec(rabbitmq@).containers
+                spec_resource::make_rabbitmq_pod_spec(rabbitmq@).containers
             );
         }
         containers
@@ -569,7 +569,7 @@ pub fn make_env_vars(rabbitmq: &RabbitmqCluster) -> (env_vars: Vec<EnvVar>)
     requires
         rabbitmq@.metadata.name.is_Some(),
     ensures
-        env_vars@.map_values(|v: EnvVar| v@) == rabbitmq_spec::make_env_vars(rabbitmq@)
+        env_vars@.map_values(|v: EnvVar| v@) == spec_resource::make_env_vars(rabbitmq@)
 {
     let mut env_vars = Vec::new();
     env_vars.push(EnvVar::new_with(
@@ -600,7 +600,7 @@ pub fn make_env_vars(rabbitmq: &RabbitmqCluster) -> (env_vars: Vec<EnvVar>)
     proof {
         assert_seqs_equal!(
             env_vars@.map_values(|v: EnvVar| v@),
-            rabbitmq_spec::make_env_vars(rabbitmq@)
+            spec_resource::make_env_vars(rabbitmq@)
         );
     }
     env_vars
