+++ src/controller_examples/rabbitmq_controller/spec/resource/mod.rs
@@ -1,2 +1,25 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
+pub mod common;
+pub mod config_map;
+pub mod default_user_secret;
+pub mod erlang_cookie;
+pub mod headless_service;
+pub mod rabbitmq_plugins;
+pub mod role;
+pub mod role_binding;
+pub mod service;
+pub mod service_account;
+pub mod stateful_set;
+
+pub use common::*;
+pub use config_map::*;
+pub use default_user_secret::*;
+pub use erlang_cookie::*;
+pub use headless_service::*;
+pub use rabbitmq_plugins::*;
+pub use role::*;
+pub use role_binding::*;
+pub use service::*;
+pub use service_account::*;
+pub use stateful_set::*;
